las
t

+++ src/vector.c
@@ -6,13 +6,13 @@
 /*   By: marlean <marlean@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/24 15:58:27 by marlean           #+#    #+#             */
-/*   Updated: 2022/07/04 13:20:07 by marlean          ###   ########.fr       */
+/*   Updated: 2022/07/05 12:51:01 by marlean          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-t_coord	new_vector3(double x, double y, double z) // Создание нового вектора
+t_coord	new_vector3(float x, float y, float z) // Создание нового вектора
 {
 	t_coord	a;
 
@@ -23,7 +23,7 @@ t_coord	new_vector3(double x, double y, double z) // Создание новог
 	return (a);
 }
 
-// t_vector2	new_vector2(double u, double v)
+// t_vector2	new_vector2(float u, float v)
 // {
 // 	t_vector2	a;
 
@@ -62,14 +62,14 @@ t_coord	vector_subtract(t_coord a, t_coord b) // Разность векторо
 // 	return (sum);
 // }
 
-double	vector_length(t_coord a) // Длина вектора
+float	vector_length(t_coord a) // Длина вектора
 {
 	return (sqrt((a.x * a.x) + (a.y * a.y) + (a.z * a.z)));
 }
 
 void	vector_normalize(t_coord *a) // Нормализация вектора
 {
-	double	length;
+	float	length;
 
 	length = vector_length(*a);
 	a->x /= length;
@@ -77,19 +77,19 @@ void	vector_normalize(t_coord *a) // Нормализация вектора
 	a->z /= length;
 }
 
-// void	vector_multiply(t_coord *direction, double multiply) // Умножение вектора на число
+// void	vector_multiply(t_coord *direction, float multiply) // Умножение вектора на число
 // {
 // 	direction->x *= multiply;
 // 	direction->y *= multiply;
 // 	direction->z *= multiply;
 // }
 
-double	vector_sumpow2(t_coord a) // Квадрат длины вектора
+float	vector_sumpow2(t_coord a) // Квадрат длины вектора
 {
 	return (pow(a.x, 2) + pow(a.y, 2) + pow(a.z, 2));
 }
 
-double	vector_scalar(t_coord a, t_coord b) // Скалярное произведение векторов
+float	vector_scalar(t_coord a, t_coord b) // Скалярное произведение векторов
 {
 	return (a.x * b.x + a.y * b.y + a.z * b.z);
 }
